ACCUMULO-1854 First stab at being able to use multiple AccumuloInputFormats and AccumuloOutputFormats in the same Configuration.
Modified the static API calls on both AIF and AOF to include the notion
of a "sequence" number. The original methods still exist, defaulting to
'0', to preserve backwards compatibility with API and functionality. A
primitive method to get a sequence number was added to both and is used
to properly generate InputSplits and output info.